Re-organize Components # Grouping Component Types in same folders
