+++ src/components/Deck/Modes/FlashCard.tsx
